[feat/#15] 캘린더 서버 연결

+++ app/src/main/java/com/aos/floney/domain/entity/CalendarItem.kt
@@ -2,7 +2,7 @@ package com.aos.floney.domain.entity
 
 data class CalendarItem(
     val date: String = "",
-    val money: Int,
+    val money: Double,
     val assetType : CalendarItemType
 )
